# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Shipment API"
  },
  "paths": {
    "/shipments/{date}": {
      "get": {
        "summary": "Retrieve all planned shipments for a date",
        "parameters": [
          {
            "in": "path",
            "name": "date",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Date to retrieve shipments for (YYYY-MM-DD)"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Shipment"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/shipments": {
      "get": {
        "summary": "Retrieve all shipments within a time period",
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "description": "Start date of the time period (inclusive)",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "description": "End date of the time period (inclusive)",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Shipment"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Shipment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "plannedDeliveryDate": {
            "type": "string",
            "format": "date-time"
          },
          "category": {
            "type": "string"
          },
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Product"
            }
          }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          }
        }
      }
    }
  }
}